Cleaned up helper functions and got cool effect in godot.
I am currently trying and failing to figure out to get a shape to keep its well, shape, when the object it is applied to changes aspect ratio. Over the course of this I have been most notably improving my UV generator function and understanding of it. I am also trying to make it so the input scale value scales the shape using it up when the value is increased and down when decreases; also failing there. But hey, failing with shaders is pretty